requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.4.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==25.9.1
gunicorn==22.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4; platform_system == 'Windows'
txaio==23.1.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
zope.interface==6.4.post2
